requirements.txt
numpy
opencv-python
Pillow
PyQt5
torch
torchvision